Move polyfills into their own file
